review: renamed favicon to signet svg and added symlink check
