+++ apps/demo-app/app/components/NavBar.vue
@@ -9,7 +9,7 @@ import {
   iconUserId,
 } from "@sit-onyx/icons";
 import type { OnyxNavBarSlots } from "sit-onyx";
-import logoUrl from "@sit-onyx/assets/logos/favicon.svg";
+import logoUrl from "@sit-onyx/assets/logos/signet.svg";
 
 defineSlots<Pick<OnyxNavBarSlots, "contextArea">>();
 const expanded = ref(false);
