+++ packages/sit-onyx/.storybook/pages/introduction.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story } from "@storybook/addon-docs/blocks";
 import { Default as ComponentShowcase } from "../../src/components/examples/ComponentShowcase/ComponentShowcase.stories.ts";
-import logo from "@sit-onyx/assets/logos/favicon.svg";
+import logo from "@sit-onyx/assets/logos/signet.svg";
 
 {/* workaround for hiding the table of contents, see docs-template.scss file */}
